cleanup major: introducing cursors
they currently have a 1-byte buffer of the incoming source code. they
simplify the stack push mechanism a little. the need to be able to push
incomplete objects on the stack (like a string during reading).

eof should be handled gracefully now. it should be easy to write a
cursor for argv.